pack.sh, *.asd: change commented-out :version lines, comment them out
pack.sh: previously, pack.sh would change a line like

  :version "$VERSION"

to

  :version "0.1.5"

(or whatever the version actually is). Now, it looks for a
commented-out line of the same code, and does the same substitution
while uncommenting the code.

This lets you keep the line with "$VERSION$", which is not a valid
version, commented out, while not running afoul of introspection tools
(e.g., uiop/utility:parse-version).

*.asd: comment out :version lines to work with this new pack.sh